requirements.txt
astroid==2.3.3
cycler==0.10.0
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.18.1
pylint==2.4.4
pymdptoolbox==4.0b3
pyparsing==2.4.6
python-dateutil==2.8.1
scipy==1.4.1
six==1.13.0
torch==1.1.0
typed-ast==1.4.1
wrapt==1.11.1
